Removed EncryptDataFile
EncryptDataFile was used to encrypt and decrypt the .rsc graphics files.
As the project is now open source, and the base graphics are available,
the encryption was simply added complexity with no benefit.

The Declare call for the EncryptDataFile from OdysseyDLL was left in
place but commented to document its existence.